% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_decompose.R
\name{bootstrap_estimate_ob_decompose}
\alias{bootstrap_estimate_ob_decompose}
\title{Bootstrapping the OB decomposition}
\usage{
bootstrap_estimate_ob_decompose(
  formula_decomposition,
  formula_reweighting,
  data_used,
  group,
  reference_0,
  normalize_factors,
  reweighting,
  reweighting_method,
  trimming,
  trimming_threshold,
  rifreg,
  rifreg_statistic,
  rifreg_probs,
  custom_rif_function,
  na.action,
  cluster = NULL,
  ...
)
}
\description{
The function resamples observations and restimates the OB decomposition
with the new sample.
}
