% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl_deco.R, R/dfl_deco_old.R
\name{dfl_deco_estimate}
\alias{dfl_deco_estimate}
\title{Estimate DFL reweighting decomposition}
\usage{
dfl_deco_estimate(
  formula,
  dep_var,
  data_used,
  weights,
  group_variable,
  reference_group,
  method,
  estimate_statistics,
  statistics,
  probs,
  reweight_marginals
)

dfl_deco_estimate(
  formula,
  dep_var,
  data_used,
  weights,
  group_variable,
  reference_group,
  method,
  estimate_statistics,
  statistics,
  probs,
  reweight_marginals
)
}
\description{
This function performs the actual DFL decomposition. It derives the
reweighting factors, estimates the distributional statistics and
calculates the decomposition terms.

This function performs the actual DFL decomposition. It derives the
reweighting factors, estimates the distributional statistics and
calculates the decomposition terms.
}
