% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl_deco_old.R, R/statistic_functions.R
\name{estimate_gini}
\alias{estimate_gini}
\title{Estimate the Gini coefficient}
\usage{
estimate_gini(dep_var, weights)

estimate_gini(dep_var, weights)
}
\description{
#' @author Rothe (2015)
estimate_gini <- function (dep_var, weights) {
  n <- length(dep_var)
  weights <- weights/sum(weights)
  G <- sum(dep_var[order(dep_var)] * 1:n * weights[order(dep_var)])
  G <- 2 * G/(n*sum(dep_var[order(dep_var)]  * weights[order(dep_var)]))
  G <- G - 1 - (1/n)
  return(G)
}
}
\author{
Rothe (2015)
}
