% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{estimate_iq_range}
\alias{estimate_iq_range}
\title{Interquantile range}
\usage{
estimate_iq_range(dep_var, weights, probs = c(0.1, 0.9))
}
\arguments{
\item{dep_var}{numeric vector of outcome variable}

\item{weights}{numeric vector of weights}

\item{probs}{a vector with probabilities whose range defines the interquantile range}
}
\value{
a numeric value indicating the (weighted) interquantile range
}
\description{
Interquantile range
}
