% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl_deco_old.R, R/statistic_functions.R
\name{estimate_iq_ratio}
\alias{estimate_iq_ratio}
\title{Interquantile ratio}
\usage{
estimate_iq_ratio(dep_var, weights, probs = c(0.1, 0.9))

estimate_iq_ratio(dep_var, weights, probs = c(0.1, 0.9))
}
\description{
Interquantile ratio

Interquantile ratio
}
