% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_deco.R
\name{estimate_ob_deco}
\alias{estimate_ob_deco}
\title{Estimate OB decomposition}
\usage{
estimate_ob_deco(
  formula,
  data_used,
  reference_0 = TRUE,
  normalize_factors = FALSE,
  compute_analytical_se = TRUE,
  return_model_fits = TRUE,
  vcov = stats::vcov
)
}
\description{
Estimate OB decomposition
}
