% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl_deco.R, R/dfl_deco_old.R
\name{fit_and_predict_probabilities}
\alias{fit_and_predict_probabilities}
\title{Predict probabilities}
\usage{
fit_and_predict_probabilities(
  formula,
  data_used,
  weights,
  method = "logit",
  newdata = NULL
)

fit_and_predict_probabilities(
  formula,
  data_used,
  weights,
  method = "logit",
  newdata = NULL
)
}
\description{
This function fits a binary choice model and predicts probabilities for every
observations.

This function fits a binary choice model and predicts probabilities for every
observations.
}
