% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{get_distributional_statistics}
\alias{get_distributional_statistics}
\title{Estimate distributional statistics}
\usage{
get_distributional_statistics(
  dep_var,
  weights,
  group_variable,
  group,
  statistics,
  probs = 1:9/10,
  log_transformed
)
}
\description{
Estimate weighted distributional statistics for the reference or
the counterfactual group.
}
