% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{get_normalized_difference}
\alias{get_normalized_difference}
\title{Get normalized differences}
\usage{
get_normalized_difference(
  formula,
  data_used,
  weights,
  psi,
  group_variable,
  reference_group
)
}
\arguments{
\item{formula}{model formula used to calulate the conditional probabilities of the reweighting factor}

\item{data_used}{\code{data.frame} with the observation for the estimation of the reweighting factor}

\item{weights}{vector with observations weights}

\item{psi}{vector with the estimated reweighting factor}

\item{group_variable}{variable with group identifier}

\item{reference_group}{identifier of (reweighted) reference group}
}
\description{
The function calculates normalized differences between covariate means of comparison
group and reweighted reference group.
}
\references{
Imbens, Guido W. and Jeffrey M. Wooldridge. 2009. Recent developments in the econometrics of program evaluation. Journal of Economic Literature 47, no. 1: 5-86.
}
