% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GU_normalization.R
\name{GU_normalization_get_coefficients}
\alias{GU_normalization_get_coefficients}
\title{Get coefficients for GU normalization}
\usage{
GU_normalization_get_coefficients(coef_names, est_coef)
}
\description{
After adjusting model.matrix and estimating the regression, this function computes
the coefficients for reference group of factors variables for GU normalization.
}
