% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GU_normalization.R
\name{GU_normalization_get_vcov}
\alias{GU_normalization_get_vcov}
\title{Sum covariance matrix for GU normalization}
\usage{
GU_normalization_get_vcov(coef_names, Cov_beta)
}
\description{
After adjusting model.matrix and estimating the regression, this function computes
the covariance matrix containing aggregate coefficients for the reference group of
GU normalized factor variables.
}
