% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GU_normalization.R
\name{GU_normalization_sum_coefficients}
\alias{GU_normalization_sum_coefficients}
\title{Sum coefficients for GU normalization}
\usage{
GU_normalization_sum_coefficients(coef_names, est_coef)
}
\description{
After adjusting model.matrix and estimating the regression, this function sums
the coefficients for reference group of GU normalized factor variables.
}
