% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GU_normalization.R
\name{GU_normalization_sum_vcov}
\alias{GU_normalization_sum_vcov}
\title{Sum covariance matrix for GU normalization}
\usage{
GU_normalization_sum_vcov(coef_names, Cov_beta)
}
\arguments{
\item{coef_names}{names of the dummy coefficients of a factor variable}

\item{Cov_beta}{estimated covariance matrix of the regression coefficients}
}
\description{
This function adjusts the covariance matrix for the additional coefficient
of the originally left-out reference level of a single factor variable.
}
