% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_functions.R
\name{integrate_generalized_lorenz_curve}
\alias{integrate_generalized_lorenz_curve}
\title{Integrate generalize Lorenz curve}
\usage{
integrate_generalized_lorenz_curve(dep_var, weights)
}
\description{
Integrate generalize Lorenz curve
}
