% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_deco.R
\name{ob_deco}
\alias{ob_deco}
\title{Oaxaca-Blinder decomposition}
\usage{
ob_deco(
  formula,
  data,
  group,
  weights = NULL,
  na.action = na.exclude,
  reference_0 = TRUE,
  normalize_factors = FALSE,
  bootstrap = FALSE,
  bootstrap_iterations = 100,
  bootstrap_robust = FALSE,
  cluster = NULL,
  cores = 1,
  vcov = stats::vcov
)
}
\arguments{
\item{formula}{an object of class "formula". See \link[stats]{lm} for further details.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{name of the a binary variable (numeric or factor)
identifying the two groups that will be compared. The group identified by the
lower ranked value in `group` (i.e., 0 in the case of a dummy variable or the
first level of factor variable) is defined as group 0.}

\item{weights}{numeric vector of non-negative observation weights, hence of same length as \code{dep_var}.
The default (\code{NULL}) is equivalent to \code{weights = rep(1, length(dep_var))}.}

\item{na.action}{generic function that defines how NAs in the data should be handled.
Default is \code{na.omit}, leading to exclusion of observations that contain one or more missings.
See \link[stats]{na.action} for further details.}

\item{reference_0}{boolean: if `TRUE` (default), then the group 0 -- i.e.,
the group identified by the lower ranked value in `group` -- will be defined
as reference group. The reference group will be reweighted to match the
covariates distribution of the counterfactual sample.}

\item{normalize_factors}{boolean: If `TRUE`, then factor variables are normalized as
proposed by Gardeazabal/Ugidos (2004) and results are not dependent on the factor's
reference group. Per default (\code{normalize_factors  = FALSE}), factors are
normalized.}

\item{bootstrap}{boolean: If `FALSE`, then the estimation is not boostrapped and no
standard errors are calculated.}

\item{bootstrap_iterations}{positive integer indicating the number of bootstrap
iterations to execute. Only required if \code{bootstrap = TRUE}.}

\item{bootstrap_robust}{boolean: if `FALSE` (default), then bootstrapped standard
errors are estimated as the standard deviations of the bootstrapp estimates.
Otherwise, the function uses the bootstrap interquartile range rescaled by the
interquantile range of the standard distribution to estimate standard errors.}

\item{cluster}{numeric vector of same length as \code{dep_var} indicating the
clustering of observations. If \code{cluster = NULL} (default), no clustering
is a assumend and bootstrap procedure resamples individual observations. Other
wise bootstrap procedure resamples clusters.}

\item{cores}{positive integer indicating the number of cores to use when
computing bootstrap standard errors. Only required if \code{bootstrap = TRUE}.}

\item{vcov}{function estimating covariance matrix of regression coefficients if
standard errors are not bootstrapped (i.e., \code{bootstrap = FALSE}). By default,
\link[stats]{vcov} is used assuming homoskedastic errors.}
}
\description{
Following Oaxaca (1973) and Blinder (1973), this function decomposes between-
group differences in the mean of an outcome variable into a part explained
by mean-differences in the observable covariates ("composition effect") and
another part due to the returns to these covariates ("structure effect"). The
function estimates the returns to covariates using OLS separately for the
two groups.
}
\examples{
mod1 <- log(wage) ~ union + married + nonwhite + education + experience
deco_results <- ob_deco(formula = mod1, data = men8305, weights = weights, group = year)
deco_results

library("AER")
data("CPS1985")
mod2 <- log(wage) ~ education + experience + union + ethnicity
deco_results2 <- ob_deco(formula = mod2, data = CPS1985, group = gender)
deco_results2

deco_results_bs <- ob_deco(formula = mod1,
                           data = men8305,
                           weights = weights,
                           group = year,
                           bootstrap = TRUE)
summary(deco_results_bs)

}
\references{
Fortin, Nicole, Thomas Lemieux, and Sergio Firpo. 2011. "Decomposition methods in economics."
In Orley Ashenfelter and David Card, eds., \emph{Handbook of labor economics}. Vol. 4. Elsevier, 1-102.

Gardeazabal, Javier, and Arantza Ugidos. 2004. "More on identification in detailed wage decompositions."
\emph{Review of Economics and Statistics} 86(4): 1034-1036.
}
