% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_deco.R
\name{ob_decompose_calculate_terms}
\alias{ob_decompose_calculate_terms}
\title{Calculate decomposition terms based on model.matrix and estimated OLS coefficients}
\usage{
ob_decompose_calculate_terms(beta0, beta1, X0, X1, weights0, weights1, reference_0)
}
\description{
Calculate decomposition terms based on model.matrix and estimated OLS coefficients
}
