% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_decompose.R
\name{ob_deco_calculate_vcov}
\alias{ob_deco_calculate_vcov}
\title{Estimate covariance matrix for OB decomposition terms}
\usage{
ob_deco_calculate_vcov(
  beta0,
  beta1,
  X0,
  X1,
  weights0,
  weights1,
  Cov_beta0,
  Cov_beta1,
  reference_0
)
}
\description{
assuming independence between groups
see also: https://www.stata.com/meeting/3german/jann.pdf
}
\details{
Var(xb) = Var(x)Var(b) + Var(x)E(b)^2 + Var(b)E(x)^2
Cov(x1b1, x2b2) = Cov(x1,x2)Cov(b1,b2) + Cov(x1,x2)E(b1)E(b2) + Cov(b1,b2)E(x1)E(x2)
Matrix notation Sigma_b * Sigma_X + Sigma_b * mu %*% mu' + Sigma_X * b %*% b'
}
