% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ob_deco}
\alias{plot.ob_deco}
\title{Plot decomposition terms for quantiles}
\usage{
\method{plot}{ob_deco}(
  x,
  confidence_bands = TRUE,
  confidence_level = 0.95,
  uniform_bands = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of class "ob_deco", usually, a result of a call to [ob_deco()] with code{statistics = "quantiles"}.}

\item{confidence_bands}{If `TRUE` (default) and if standard errors have been bootstrapped, confidence bands are plotted.}

\item{confidence_level}{numeric value between 0 and 1 (default = 0.95) that defines the confidence interval
plotted as a ribbon and defined as \code{qnorm((1-confidence_level)/2)} * standard error.}

\item{uniform_bands}{If `FALSE` (default), pointsise confidence bands are computed. Otherwise, uniform bands are constructed
based on the bootstrapped Kolmogrov-Smirnov statistic.}

\item{...}{other parameters to be passed through to plotting functions.}
}
\value{
a ggplot illustrating the decomposition terms for quantiles.
}
\description{
The function plots decomposition terms for quantiles estimtated
with \code{ob_deco} over the  unit interval.
}
\examples{
# NOT YET PROVIDED

}
