% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dfl_deco}
\alias{print.dfl_deco}
\title{print method for class "dfl_deco"}
\usage{
\method{print}{dfl_deco}(x, ...)
}
\arguments{
\item{x}{an object of class "dfl_deco", usually , a result of a call to [dfl_deco()].}

\item{...}{other parameters to be passed through to printing functions.}
}
\value{
The function \code{print.dfl_deco()} displays the decompositions terms saved in \code{x}.
}
\description{
print method for class "dfl_deco"
}
