% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dfl_decompose}
\alias{print.dfl_decompose}
\title{print method for class "dfl_decompose"}
\usage{
\method{print}{dfl_decompose}(x, ...)
}
\arguments{
\item{x}{an object of class "dfl_decompose", usually , a result of a call to [dfl_decompose()].}

\item{...}{other parameters to be passed through to printing functions.}
}
\value{
The function \code{print.dfl_decompose()} displays the decompositions terms saved in \code{x}.
}
\description{
print method for class "dfl_decompose"
}
