% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ob_deco}
\alias{print.ob_deco}
\title{print method for class "ob_deco"}
\usage{
\method{print}{ob_deco}(x, ...)
}
\arguments{
\item{x}{an object of class "ob_deco", usually , a result of a call to [ob_deco()].}

\item{...}{other parameters to be passed through to printing functions.}
}
\value{
The function \code{print.ob_deco()} displays the decompositions terms saved in \code{x}.
}
\description{
print method for class "ob_deco"
}
