% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ob_decompose}
\alias{print.ob_decompose}
\title{print method for class "ob_decompose"}
\usage{
\method{print}{ob_decompose}(x, ...)
}
\arguments{
\item{x}{an object of class "ob_decompose", usually , a result of a call to [ob_decompose()].}

\item{...}{other parameters to be passed through to printing functions.}
}
\value{
The function \code{print.ob_decompose()} displays the decompositions terms saved in \code{x}.
}
\description{
print method for class "ob_decompose"
}
