% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl_deco_old.R, R/summary.R
\name{summary.dfl_deco}
\alias{summary.dfl_deco}
\title{summary method for class "dfl_deco"}
\usage{
\method{summary}{dfl_deco}(x, confidence_level = 0.95, digits = 4, ...)

\method{summary}{dfl_deco}(x, confidence_level = 0.95, digits = 4, ...)
}
\arguments{
\item{x}{an object of class "dfl_deco", a result of a call to [dfl_deco()].}

\item{confidence_level}{numeric value between 0 and 1 (default = 0.95) that defines the
confidence level of the printed confidence intervals. Pointwise confidences bands
are defined as \code{qnorm((1-confidence_level)/2)} * standard error. Uniform bands
are constructed by multiplying the standard error with \code{confidence_level}-quantile 
of the bootstrapped Kolmogorov-Smirnov statistic.}

\item{...}{other parameters to be passed through to printing functions.}
}
\value{
The function \code{summary.dfl_deco()} displays the decompositions
terms save in \code{x}. If standard errors have been bootstrapped, standard
errors and confidence bands are given. Uniform confidence bands for quantiles
are constructed based on the bootstrappend Kolmogorov-Smirnov distribution.

The function \code{summary.dfl_deco()} displays the decompositions 
terms save in \code{x}. If standard errors have been bootstrapped, standard 
errors and confidence bands are given. Uniform confidence bands for quantiles
are constructed based on the bootstrappend Kolmogorov-Smirnov distribution.
}
\description{
summary method for class "dfl_deco"

summary method for class "dfl_deco"
}
