% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ob_deco}
\alias{summary.ob_deco}
\title{summary method for class "ob_decompose"}
\usage{
\method{summary}{ob_deco}(
  x,
  aggregate_factors = TRUE,
  custom_aggregation = NULL,
  confidence_level = 0.95,
  ...
)

\method{summary}{ob_deco}(
  x,
  aggregate_factors = TRUE,
  custom_aggregation = NULL,
  confidence_level = 0.95,
  ...
)
}
\arguments{
\item{x}{an object of class "ob_decompose", usually , a result of a call to [ob_decompose()].}

\item{aggregate_factors}{boolean, if `TRUE` (default) terms associated with detailed factor
levels are aggregated to a single term for every factor variable.}

\item{custom_aggregation}{list specifying the aggregation of detailed decomposition
terms. The parameter `custom_aggregation` overrides the parameter `aggregate_factors`.
If `NULL` (default), then either all detailed terms or all terms associated with
a single variable are returned.}

\item{confidence_level}{numeric value between 0 and 1 (default = 0.95) that defines the printed confidence interval.}

\item{...}{other parameters to be passed through to print functions.}
}
\value{
The function \code{summary.ob_decompose()} summarizes the decompositions terms saved in \code{x}.

The function \code{summary.ob_decompose()} summarizes the decompositions terms saved in \code{x}.
}
\description{
Apart from displaying the (detailed) decomposition results with standard
errors, \code{summary.ob_decompose()} allows to customize the aggregation of the
detailed decomposition terms.

Apart from displaying the (detailed) decomposition results with standard
errors, \code{summary.ob_decompose()} allows to customize the aggregation of the
detailed decomposition terms.
}
\examples{
library("AER")
data("CPS1985")
mod2 <- log(wage) ~ education + experience + union + ethnicity
deco_results <- ob_decompose(formula = mod2, data = CPS1985, group = gender)

# Print standard errors
summary(deco_results)

# Aggregate decomposition terms associated with factor levels
summary(deco_results, aggregate_factors = TRUE)

# custom aggregation of decompisition terms
custom_aggregation <- list(human_capital = c("education", "experience"),
                           union = "unionyes")
summary(deco_results, custom_aggregation = custom_aggregation)

library("AER")
data("CPS1985")
mod2 <- log(wage) ~ education + experience + union + ethnicity
deco_results <- ob_decompose(formula = mod2, data = CPS1985, group = gender)

# Print standard errors
summary(deco_results)

# Aggregate decomposition terms associated with factor levels
summary(deco_results, aggregate_factors = TRUE)

# custom aggregation of decompisition terms
custom_aggregation <- list(human_capital = c("education", "experience"),
                           union = "unionyes")
summary(deco_results, custom_aggregation = custom_aggregation)

}
