% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{aggregate_terms}
\alias{aggregate_terms}
\title{Aggregate decomposition terms}
\usage{
aggregate_terms(x, aggregate_factors = TRUE, custom_aggregation = NULL)

aggregate_terms(x, aggregate_factors = TRUE, custom_aggregation = NULL)
}
\arguments{
\item{x}{an object of class "ob_deco", usually , a result of a call to [ob_deco()].}

\item{aggregate_factors}{boolean, if `TRUE` (default) terms associated with detailed factor
levels are aggregated to a single term for every factor variable.}

\item{custom_aggregation}{list specifying the aggregation of detailed decomposition
terms. The parameter `custom_aggregation` overrides the parameter `aggregate_factors`.
If `NULL` (default), then either all detailed terms or all terms associated with
a single variable are returned.}
}
\value{
The function returns an updated object of class "ob_deco" containing
the aggregated decomposition terms.

The function returns an updated object of class "ob_deco" containing
the aggregated decomposition terms.
}
\description{
The function aggregates decomposition terms and calculates
their covariance matrix based on detailed decomposition results.

The function aggregates decomposition terms and calculates
their covariance matrix based on detailed decomposition results.
}
\examples{
# library("AER")
# data("CPS1985")
# mod2 <- log(wage) ~ education + experience + union + ethnicity
# deco_results <- ob_deco(formula = mod2, data = CPS1985, group = gender)
#
# custom_aggregation <- list(human_capital = c("education", "experience"),
#                            union = "unionyes")
# aggregated_results <- aggregate_terms(deco_results2,
#                                       custom_aggregation = custom_aggregation)

}
