% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_deco.R
\name{bootstrap_estimate_ob_deco}
\alias{bootstrap_estimate_ob_deco}
\title{Estimate OB decomposition in bootstrap replications}
\usage{
bootstrap_estimate_ob_deco(
  formula,
  data_used,
  reference_0,
  normalize_factors,
  reweighting,
  reweighting_method,
  rifreg,
  cluster = NULL
)
}
\description{
Estimate OB decomposition in bootstrap replications
}
