% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_deco.R
\name{bootstrap_estimate_ob_deco}
\alias{bootstrap_estimate_ob_deco}
\title{Estimate OB decomposition in bootstrap replications}
\usage{
bootstrap_estimate_ob_decompose(
  formula,
  data_used,
  reference_0 = TRUE,
  normalize_factors = FALSE,
  cluster = NULL
)
}
\description{
Estimate OB decomposition in bootstrap replications
}
