% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl_deco.R
\name{dfl_deco_bootstrap}
\alias{dfl_deco_bootstrap}
\title{Bootstrap function}
\usage{
dfl_deco_bootstrap(
  formula,
  dep_var,
  data_used,
  weights,
  group_variable,
  reference_group,
  estimate_statistics,
  statistics,
  probs,
  reweight_marginals,
  ...
)
}
\description{
Bootstrap function
}
