% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl_decomposeR
\name{dfl_decompose_estimate}
\alias{dfl_decompose_estimate}
\title{Estimate DFL reweighting decomposition}
\usage{
dfl_decompose_estimate(
  formula,
  dep_var,
  data_used,
  weights,
  group_variable,
  reference_group,
  method,
  estimate_statistics,
  statistics,
  probs,
  reweight_marginals
)
}
\description{
This function performs the actual DFL decomposition. It derives the
reweighting factors, estimates the distributional statistics and
calculates the decomposition terms.
}
