% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfl_decompose.R
\name{fit_and_predict_probabilities}
\alias{fit_and_predict_probabilities}
\title{Predict conditional probabilities}
\usage{
fit_and_predict_probabilities(
  formula,
  data_used,
  weights,
  method = "logit",
  return_model = FALSE,
  newdata = NULL,
  ...
)
}
\arguments{
\item{formula}{\code{formula} object specifying the conditional probability model.}

\item{data_used}{\code{data.frame} with data.}

\item{weights}{weights variable}

\item{method}{method to estimate conditional probabilities}

\item{return_model}{boolean: If \code{FALSE} (default), the object of the model
fit used to predict the conditional probabilities for the reweighting factor
are not returned.}

\item{newdata}{\code{data.frame} with data to be used for predictions.}

\item{...}{other parameters passed to the estimation function.}
}
\description{
This function fits a binary choice model and predicts probabilities for every
observations.
}
