% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GU_normalization.R
\name{GU_normalization_get_coefficients}
\alias{GU_normalization_get_coefficients}
\title{Get coefficients for GU normalization}
\usage{
GU_normalization_get_coefficients(coef_names, est_coef)
}
\arguments{
\item{coef_names}{list with coefficients of every factor variable that need to be adjusted}

\item{est_coef}{vector of estimated coefficients}
}
\description{
This function constructs sums the coefficients of each factor variable to
construct a additional coefficients for their originally left-out reference
levels and adds them to the estimated coefficients vector.
}
