% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GU_normalization.R
\name{GU_normalization_sum_coefficients}
\alias{GU_normalization_sum_coefficients}
\title{Sum coefficients for GU normalization}
\usage{
GU_normalization_sum_coefficients(coef_names, est_coef)
}
\arguments{
\item{coef_names}{names of the dummy coefficients of a factor variable}

\item{est_coef}{estimated coefficient vector}
}
\description{
This function sums the coefficients of a single factor variable to construct
an additional coefficient for the left-out reference level.
}
