% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{men8305}
\alias{men8305}
\title{Sample of male wage data from the CPS 1983-1985 and 2003-2005}
\format{
A data frame with 40,347 rows and 8 variables.
\describe{
  \item{wage}{Hourly wage in US dollars at constant prices}
  \item{union}{Union status indicator}
  \item{education}{Factor variable with 6 education levels: high-school graduates (reference), elementary, high-school dropouts , some college, college graduates, post college graduates}
  \item{experience}{Factor variable with 9 potential experience levels, each of five years gap, 20 to 24 years as reference level)}
  \item{married}{Married indicator}
  \item{nonwhite}{Non-white indicator}
  \item{year}{Indicator distinguishing pooled observations from the 1983 to 1985 period and those from 2003 to 2005}
  \item{weights}{CPS sample weights}
}
}
\source{
Fortin, Nicole M., Thomas Lemieux, and Firpo Segio. 2011.
"Decomposition Methods in Economics." In Orley Ashenfelter and David Card, eds.,
Handbook of Labor Economics, Volume 4a., Chapter 1, 1-102.
}
\usage{
men8305
}
\description{
A sample of the the Merged Outgoing Rotation Group of the
Current Population Survey from 1983 to 1985 and 2003 to 2005, respectively,
used as example by Fortin, Lemieux & Firpo  (2011) in their handbook chapter.
The data set contains a selection of 8 variables and a sample of 40,347 observations
of male workers (i.e., a tenth of the origninal data set).
}
\keyword{datasets}
