% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_deco.R
\name{ob_deco}
\alias{ob_deco}
\title{Oaxaca-Blinder decomposition}
\usage{
ob_deco(
  formula,
  data,
  group,
  swap = FALSE,
  weights = NULL,
  rifreg = FALSE,
  rifreg_statistic = "quantiles",
  rifreg_probs = c(1:9)/10,
  custom_rif_function = NULL,
  reweighting = FALSE,
  reweighting_method = "logit",
  reference_0 = TRUE,
  na.action = na.omit,
  normalize_factors = FALSE,
  bootstrap = FALSE,
  bootstrap_iterations = 100,
  bootstrap_robust = FALSE,
  cluster = NULL,
  cores = 1,
  vcov = stats::vcov,
  ...
)
}
\arguments{
\item{formula}{an object of class "formula". See \link[stats]{lm} for further details.}

\item{data}{a data frame containing the variables in the model.}

\item{group}{name of the a binary variable (numeric or factor)
identifying the two groups that will be compared. The group identified by the
lower ranked value in `group` (i.e., 0 in the case of a dummy variable or the
first level of factor variable) is defined as group 0.}

\item{swap}{boolean: `FALSE` is default. Setting it to `TRUE` merely changes the sign of the decomposition results.
This means the composition effect is computed as \code{(X0 - X1) * b1} and
the structure effect as \code{X0 * (b0 - b1)}.}

\item{weights}{numeric vector of non-negative observation weights, hence of same length as \code{dep_var}.
The default (\code{NULL}) is equivalent to \code{weights = rep(1, length(dep_var))}.}

\item{rifreg}{boolean: if `TRUE`, then a RIF regression for the statistic
indicated in `rifreg_statistic` is computed and decomposed.}

\item{rifreg_statistic}{string containing the distributional statistic for which to compute the RIF. Can be one of
"quantiles", "mean", "variance", "gini", "interquantile_range", "interquantile_ratio", or "custom".
Default is "quantiles". If "custom" is selected, a \code{custom_rif_function} needs to be provided.}

\item{rifreg_probs}{a vector of length 1 or more with probabilities of quantiles. Each quantile is indicated with a value between 0 and 1.
Default is \code{c(1:9)/10}. If \code{statistic = "quantiles"}, a single RIF regression for every quantile in \code{probs}
is estimated. An interquantile ratio (range) is defined by the ratio (difference) between the \code{max(probs)}-quantile and
the \code{min(probs)}-quantile.}

\item{custom_rif_function}{the RIF function to compute the RIF of the custom distributional statistic.
Default is NULL. Only needs to be provided if \code{statistic = "custom"}.
Every custom_rif_function needs the parameters \code{dep_var}, \code{weights} and \code{probs}.
If they are not needed, they must be set to NULL in the function definition (e.g. \code{probs = NULL}).
A custom function must return a data frame containing at least a "rif" and "weights" column.
See \code{examples} for further details.}

\item{reweighting}{boolean: if `TRUE`, then reweighted RIF regressions are computed.}

\item{reference_0}{boolean: if `TRUE` (default), then the group 0 -- i.e.,
the group identified by the lower ranked value in `group` -- will be defined
as reference group. The reference group will be reweighted to match the
covariates distribution of the counterfactual sample.
By default, the composition effect is computed as \code{(X1 - X0) * b0} and
the structure effect as \code{X1 * (b1 - b0)}. Putting \code{reference_0 == FALSE} changes
the reference structure. Hence the composition effect is computed as \code{(X1 - X0) * b1} and
the structure effect as \code{X0 * (b1 - b0)}.}

\item{na.action}{generic function that defines how NAs in the data should be handled.
Default is \code{na.omit}, leading to exclusion of observations that contain one or more missings.
See \link[stats]{na.action} for further details.}

\item{normalize_factors}{boolean: If `TRUE`, then factor variables are normalized as
proposed by Gardeazabal/Ugidos (2004) and results are not dependent on the factor's
reference group. Per default (\code{normalize_factors  = FALSE}), factors are
normalized.}

\item{bootstrap}{boolean: If `FALSE`, then the estimation is not boostrapped and no
standard errors are calculated.}

\item{bootstrap_iterations}{positive integer indicating the number of bootstrap
iterations to execute. Only required if \code{bootstrap = TRUE}.}

\item{bootstrap_robust}{boolean: if `FALSE` (default), then bootstrapped standard
errors are estimated as the standard deviations of the bootstrapp estimates.
Otherwise, the function uses the bootstrap interquartile range rescaled by the
interquantile range of the standard distribution to estimate standard errors.}

\item{cluster}{numeric vector of same length as \code{dep_var} indicating the
clustering of observations. If \code{cluster = NULL} (default), no clustering
is a assumend and bootstrap procedure resamples individual observations. Other
wise bootstrap procedure resamples clusters.}

\item{cores}{positive integer indicating the number of cores to use when
computing bootstrap standard errors. Only required if \code{bootstrap = TRUE}.}

\item{vcov}{function estimating covariance matrix of regression coefficients if
standard errors are not bootstrapped (i.e., \code{bootstrap = FALSE}). By default,
\link[stats]{vcov} is used assuming homoskedastic errors.}

\item{...}{additional parameters passed to the custom_rif_function.
Apart from dep_var, weights and probs they must have a different name than the the ones in rifreg.
For instance, if you want to pass a parameter statistic to the custom_rif_function, name it custom_statistic.
Additional parameters can also be passed to the \link[stats]{density} function.}
}
\description{
Following Oaxaca (1973) and Blinder (1973), this function decomposes between-
group differences in the mean of an outcome variable into a part explained
by mean-differences in the observable covariates ("composition effect") and
another part due to the returns to these covariates ("structure effect"). The
function estimates the returns to covariates using OLS separately for the
two groups.
}
\examples{

## Decompose gender wage gap
## with NLYS79 data like in Fortin, Lemieux, & Firpo (2011: 41)

data("nlys00")

mod1 <- log(wage) ~ age + central_city + msa + region + black +
  hispanic + education + afqt + family_responsibility + years_worked_civilian +
  years_worked_military + part_time + industry

# Using female coefficients (reference_0 = TRUE) to estimate counterfactual mean
deco_female_as_reference <- ob_deco(formula = mod1,
                                    data = nlys00,
                                    group = female,
                                    reference_0 = TRUE)
deco_female_as_reference

# Using male coefficients (reference_0 = FALSE)
deco_male_as_reference <- ob_deco(formula = mod1,
                                  data = nlys00,
                                  group = female,
                                  reference_0 = FALSE)
deco_male_as_reference

# Replicate first and third column in Table 3 in Fortin, Lemieux, & Firpo (2011: 41)
# Define aggregation of decomposition terms
custom_aggregation <- list(`Age, race, region, etc.` = c("age",
                                                         "blackyes",
                                                         "hispanicyes",
                                                         "regionNorth-central",
                                                         "regionSouth",
                                                         "regionWest",
                                                         "central_cityyes",
                                                         "msayes"),
                   `Education` = c("education<10 yrs",
                                   "educationHS grad (diploma)",
                                   "educationHS grad (GED)",
                                   "educationSome college",
                                   "educationBA or equiv. degree",
                                   "educationMA or equiv. degree",
                                   "educationPh.D or prof. degree"),
                   `AFTQ` = "afqt",
                   `L.T. withdrawal due to family` =  "family_responsibility",
                   `Life-time work experience` = c("years_worked_civilian",
                                                   "years_worked_military",
                                                   "part_time"),
                   `Industrial sectors` = c("industryManufacturing",
                                            "industryEducation, Health, Public Admin.",
                                            "industryOther services"))

# First column
summary(deco_male_as_reference, custom_aggregation = custom_aggregation)

# Third column
summary(deco_female_as_reference, custom_aggregation = custom_aggregation)

## Compare bootstrapped standard errors...
deco_female_as_reference_bs <- ob_deco(formula = mod1,
                                       data = nlys00,
                                       group = female,
                                       bootstrap = TRUE,
                                       bootstrap_iterations = 100)
summary(deco_female_as_reference_bs, custom_aggregation = custom_aggregation)

# ... to analytical standard errors (assuming independence between groups and
# homoskedasticity)
deco_female_as_reference <- ob_deco(formula = mod1,
                                    data = nlys00,
                                    group = female,
                                    reference_0 = TRUE)
summary(deco_female_as_reference, custom_aggregation = custom_aggregation)

# Return standard errors for all detailed terms -> not yet working
# summary(deco_female_as_reference, aggregate_factors = FALSE)

}
\references{
Fortin, Nicole, Thomas Lemieux, and Sergio Firpo. 2011. "Decomposition methods in economics."
In Orley Ashenfelter and David Card, eds., \emph{Handbook of labor economics}. Vol. 4. Elsevier, 1-102.

Gardeazabal, Javier, and Arantza Ugidos. 2004. "More on identification in detailed wage decompositions."
\emph{Review of Economics and Statistics} 86(4): 1034-1036.
}
