% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_deco.R
\name{ob_deco_calculate_vcov}
\alias{ob_deco_calculate_vcov}
\title{Estimate covariance matrix for OB decomposition terms}
\usage{
ob_deco_calculate_vcov(
  beta0,
  beta1,
  X0,
  X1,
  weights0,
  weights1,
  Cov_beta0,
  Cov_beta1,
  reference_0 = TRUE
)
}
\description{
assuming indenpendence between groups
see: https://www.stata.com/meeting/3german/jann.pdf
}
