% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ob_deco}
\alias{plot.ob_deco}
\title{Plot decomposition terms for quantiles}
\usage{
\method{plot}{ob_deco}(x, confidence_bands = TRUE, confidence_level = 0.95)
}
\arguments{
\item{x}{an object of class "ob_deco", usually, a result of a call to [ob_deco()] with code{statistics = "quantiles"}.}

\item{confidence_bands}{If `TRUE` (default) and if standard errors have been bootstrapped, confidence bands are plotted.}

\item{confidence_level}{numeric value between 0 and 1 (default = 0.95) that defines the confidence interval
plotted as a ribbon and defined as \code{qnorm((1-confidence_level)/2)} * standard error.}
}
\value{
a ggplot illustrating the decomposition terms for quantiles.
}
\description{
The function plots decomposition terms for quantiles estimtated
with \code{ob_deco} over the  unit interval.
}
\examples{
data("nlys00")

mod1 <- log(wage) ~ age + central_city + msa + region + black +
  hispanic + education + afqt + family_responsibility + years_worked_civilian +
  years_worked_military + part_time + industry

deco_male_as_reference <- ob_deco(formula = mod1,
                                  data = nlys00,
                                  group = female,
                                  reweighting = TRUE,
                                  rifreg = TRUE,
                                  bootstrap = TRUE,
                                  bootstrap_iterations = 50,
                                  reference_0 = FALSE)

plot(deco_male_as_reference)

plot(deco_male_as_reference,
     confidence_bands = FALSE)

}
